+++ service/controller/clusterapi/v29/resource/ipam/cluster_checker.go
@@ -1,4 +1,4 @@
-package network
+package ipam
 
 import (
 	"context"
