+++ service/controller/clusterapi/v29/resource/ipam/subnet_collector.go
@@ -1,4 +1,4 @@
-package network
+package ipam
 
 import (
 	"context"
